Move *html.Node tools into xhtml; merge inline siblings
