+++ nkotb/pkg/blocko/blocko.go
@@ -6,6 +6,7 @@ import (
 	"io"
 	"strings"
 
+	"github.com/spotlightpa/nkotb/pkg/xhtml"
 	"golang.org/x/net/html"
 	"golang.org/x/net/html/atom"
 )
@@ -36,14 +37,14 @@ loop:
 				fmt.Fprint(w, strings.Repeat("    ", depth))
 			}
 		}
-		if !blockElements[p.DataAtom] {
+		if !xhtml.BlockElements[p.DataAtom] {
 			if err = html.Render(&wbuf, p); err != nil {
 				return err
 			}
 			needsNL = output(w, &wbuf)
 			continue
 		}
-		if IsEmpty(p) {
+		if xhtml.IsEmpty(p) {
 			fmt.Fprint(w, "")
 			needsNL = false
 			continue
@@ -86,94 +87,4 @@ func output(w io.Writer, wbuf *strings.Builder) bool {
 	wbuf.Reset()
 	fmt.Fprint(w, s)
 	return s != ""
-}
-
-func findNode(n *html.Node, callback func(*html.Node) *html.Node) *html.Node {
-	if r := callback(n); r != nil {
-		return r
-	}
-	for c := n.FirstChild; c != nil; c = c.NextSibling {
-		if r := findNode(c, callback); r != nil {
-			return r
-		}
-	}
-	return nil
-}
-
-func visitAll(n *html.Node, callback func(*html.Node)) {
-	findNode(n, func(n *html.Node) *html.Node {
-		callback(n)
-		return nil
-	})
-}
-
-var blockElements = map[atom.Atom]bool{
-	atom.P:  true,
-	atom.H1: true,
-	atom.H2: true,
-	atom.H3: true,
-	atom.H4: true,
-	atom.H5: true,
-	atom.H6: true,
-	atom.Ul: true,
-	atom.Ol: true,
-}
-
-var stylisticElements = map[atom.Atom]bool{
-	atom.A:       true,
-	atom.Abbr:    true,
-	atom.Acronym: true,
-	atom.B:       true,
-	atom.Bdi:     true,
-	atom.Bdo:     true,
-	atom.Big:     true,
-	atom.Br:      true,
-	atom.Cite:    true,
-	atom.Code:    true,
-	atom.Del:     true,
-	atom.Dfn:     true,
-	atom.Em:      true,
-	atom.I:       true,
-	atom.Ins:     true,
-	atom.Kbd:     true,
-	atom.Label:   true,
-	atom.Mark:    true,
-	atom.Meter:   true,
-	atom.Output:  true,
-	atom.Q:       true,
-	atom.Ruby:    true,
-	atom.S:       true,
-	atom.Samp:    true,
-	atom.Small:   true,
-	atom.Span:    true,
-	atom.Strong:  true,
-	atom.Sub:     true,
-	atom.Sup:     true,
-	atom.U:       true,
-	atom.Tt:      true,
-	atom.Var:     true,
-	atom.Wbr:     true,
-}
-
-func IsEmpty(n *html.Node) bool {
-	root := n
-	n = findNode(n, func(n *html.Node) *html.Node {
-		if n == root {
-			return nil
-		}
-		switch n.Type {
-		case html.TextNode:
-			s := strings.ReplaceAll(n.Data, "\n", " ")
-			s = strings.TrimSpace(s)
-			if s == "" {
-				return nil
-			}
-		case html.ElementNode:
-			if stylisticElements[n.DataAtom] {
-				return nil
-			}
-		}
-		return n
-	})
-	return n == nil
 }
