+++ nkotb/pkg/blocko/minify.go
@@ -5,13 +5,15 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/spotlightpa/nkotb/pkg/xhtml"
 	"github.com/tdewolff/minify/v2"
 	"github.com/tdewolff/minify/v2/html"
-	xhtml "golang.org/x/net/html"
+	"golang.org/x/exp/slices"
+	nethtml "golang.org/x/net/html"
 	"golang.org/x/net/html/atom"
 )
 
-func prep(r io.Reader) (*xhtml.Node, error) {
+func prep(r io.Reader) (*nethtml.Node, error) {
 	m := minify.New()
 	m.AddFunc("text/html", html.Minify)
 
@@ -19,11 +21,11 @@ func prep(r io.Reader) (*xhtml.Node, error) {
 	if err := m.Minify("text/html", &out, r); err != nil {
 		return nil, err
 	}
-	doc, err := xhtml.Parse(&out)
+	doc, err := nethtml.Parse(&out)
 	if err != nil {
 		return nil, err
 	}
-	body := findNode(doc, func(n *xhtml.Node) *xhtml.Node {
+	body := xhtml.Find(doc, func(n *nethtml.Node) *nethtml.Node {
 		if n.DataAtom == atom.Body {
 			return n
 		}
@@ -32,5 +34,32 @@ func prep(r io.Reader) (*xhtml.Node, error) {
 	if body == nil {
 		return nil, fmt.Errorf("could not find body")
 	}
+
+	mergeSiblings(body)
+
 	return body, nil
+}
+
+func mergeSiblings(root *nethtml.Node) {
+	// find all matches first
+	var inlineSiblings []*nethtml.Node
+	xhtml.VisitAll(root, func(n *nethtml.Node) {
+		brother := n.NextSibling
+		if brother == nil ||
+			!xhtml.InlineElements[n.DataAtom] ||
+			n.DataAtom != brother.DataAtom ||
+			!slices.Equal(n.Attr, brother.Attr) {
+			return
+		}
+		inlineSiblings = append(inlineSiblings, n)
+	})
+	// then do mutation.
+	// no mutating while iterating!
+	// go in reverse order
+	// in case there are several siblings to merge
+	for i := len(inlineSiblings) - 1; i >= 0; i-- {
+		n := inlineSiblings[i]
+		xhtml.AdoptChildren(n, n.NextSibling)
+		n.Parent.RemoveChild(n.NextSibling)
+	}
 }
