+++ nkotb/pkg/blocko/testdata/siblings.html
@@ -1,0 +1,16 @@
+<p>
+    <b>Hello,</b><b> world!</b>.
+    <b>One,</b><b> two,</b><b> three</b>.
+</p>
+<p>
+    <a href="one">Link one</a><a href="one"> and one</a>
+</p>
+<p>
+    <a href="one">Link 1</a><a href="two">Link 2</a>
+</p>
+<div>
+    <em>Don't touch that div!</em>
+</div>
+<div>
+    <em>Do not touch that div!</em>
+</div>
