+++ nkotb/pkg/blocko/testdata/siblings.md
@@ -1,0 +1,10 @@
+<b>Hello, world!</b>.
+<b>One, two, three</b>.
+
+<a href="one">Link one and one</a>
+
+<a href="one">Link 1</a><a href="two">Link 2</a>
+
+<div><em>Don&#39;t touch that div!</em></div>
+
+<div><em>Do not touch that div!</em></div>
